+++ general/models.py
@@ -1,10 +1,12 @@
+import datetime
+
 from django.db import models
 
 
 class AbstractEntity(models.Model):
     id = models.AutoField(primary_key=True)
-    created = models.DateTimeField(auto_created=True)
-    updated = models.DateTimeField(auto_now_add=True)
+    created = models.DateTimeField(auto_now_add=True, null=True, blank=True)
+    updated = models.DateTimeField(auto_now_add=True, null=True, blank=True)
 
     class Meta:
         abstract = True
